Correctly detect source dir and execute prereqs script; only install .local_home symlink if not exists

+++ install.sh
@@ -2,13 +2,15 @@
 
 set -x
 
+SOURCE_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
 cd ~/
 
 if [ ! -f ~/.ssh/id_rsa.pub ]; then
 	ssh-keygen
 fi
 
-./prereqs.sh
+exec $SOURCE_DIR/prereqs.sh
 
 
 if [ ! -d ~/.rbenv ]; then
@@ -18,8 +20,7 @@ if [ ! -d ~/.rbenv ]; then
     git clone https://github.com/rbenv/ruby-build.git ~/.rbenv/plugins/ruby-build
 fi
 
-
-SOURCE_DIR="$( dirname "$0")"
-
-ln -s $SOURCE_DIR ~/.local_home
-echo ". ~/.local_home/.rc" >> ~/.bashrc
+if [ ! -e ~/.local_home ]; then
+	ln -s $SOURCE_DIR ~/.local_home
+	echo ". ~/.local_home/.rc" >> ~/.bashrc
+fi
